added calculateAccerlation implementations (WIP) and clipping
